backend/requirements.txt
fastapi
uvicorn[standard]
httpx
apscheduler
anthropic
python-dotenv
aiosqlite
numpy
pandas
